314 ORM compatibility for CDL:
General cleanup:  Quit crashing all of gips if the user hasn't put in
their CSV file.  Use one module attrib for all the special driver
strings (product type, asset type, sensor).  Refactor asset filename
parser from daymet to be available to all drivers.

ORM compat: Same as recent work for daymet:  We name the assets however
we want, so name them so they look like products, and thus we get
--rectify for free.
